backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
bidict==0.23.1
cachetools==5.5.2
certifi==2025.1.31
channels==4.2.2
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
coverage==7.8.0
decorator==5.2.1
distro==1.9.0
django==5.2
django-cors-headers==4.7.0
django-mongodb-backend==5.1.0b1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
executing==2.2.0
filelock==3.18.0
fsspec==2025.3.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
jedi==0.19.2
jiter==0.9.0
jupyterlab_widgets==3.0.14
matplotlib-inline==0.1.7
mutagen==1.47.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.12.0
pyparsing==3.2.3
pytest==8.3.5
pytest-cov==6.1.1
python-dotenv==1.1.0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
simple-websocket==1.1.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
widgetsnbextension==4.0.14
wsproto==1.2.0
